Small refactor. Add `__init__.py`
